Add choose_game()
choose_game() gets a list of saved game worksheets,
It displays the list with indexes,
It invites the player to choose a game ot review,
or to continue on to playing the game.

In this commit, Games._duration was initiated as 0.0,
final_update() in game.py was amended to write delta.total_seonds()
to game worksheet so that game duration could be compared later.
display_game_data() in game.py was modified to convert
self._duration useble form:
        delta_min_sec = divmod(float(self._duration), 60)
        mins = int(delta_min_sec[0])
        secs = int(delta_min_sec[1])
        print(f'The game lasted {mins}m {secs}s')